requirements.txt
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
Pillow==12.0.0
scipy==1.16.3
skimage==0.0
tqdm==4.65.0
